backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
python-dotenv>=1.0.0
pydantic>=2.5.0
langchain-google-genai>=1.0.0
langgraph>=0.2.0
google-genai>=0.3.0
networkx>=3.0
chromadb>=0.3.21
structlog>=24.0.0
prometheus_client>=0.20.0
